+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### [Carp-0.#.#-alpha]()
 - Added an argument for pretty printing
 - Fixed [#22](https://github.com/DeflatedPickle/Carp/issues/22)
+- Added function parameters
+- Fixed function parameters
+- Added return statements
 
 ### [Carp-0.6.1-alpha](https://github.com/DeflatedPickle/Carp/releases/tag/v0.6.1-alpha)
 - Added classes